requirements.txt
Flask~=2.2.3
python-telegram-bot~=20.1
gunicorn~=20.1.0